refactor: rename default data file
