+++ front/src/pages/people/default-data.ts
